Fixes incorrect variant images being displayed in stock management
Currently when a variant is supplied to one of the generated image helpers and the variant does not have any associated images, the first of the variant’s product’s variant_images is used. The result is that incorrect images are displayed for certain variants. This can be misleading to admins since the image isn’t incorrect, it’s just not present.

This change is the initial step to begin deprecating the image helper methods. The image helpers’ behavior isn’t very clear so we want to move towards deprecating them in favor of accessing the images associated with a resource directly. A deprecation warning was added to the currently generated helper methods.