Increment stock when an item is returned to a location for which it does not yet have stock.
Prior to this code, if an item is returned to a stock location that it
does not yet have a stock item for, it would not increment stock at that
stock location upon being received. This change creates a stock item for
the returned variant for the stock location if it doesn't yet exist.