Don't re-authorize already authorized payments
cherry-picked from https://github.com/jordan-brough/solidus/commit/736822a

If a payment has already been authorized we should not re-authorize
it during Payment#process.  This can happen when an order has multiple
payments and one of them fails, sending the order back to a
pre-complete state.  On the next attempt to complete the order,
'#process' will be called on the already-authorized payment again.
This results in errors with store credits and in over-authorization
with credit cards.